+++ src/scripts/interfaces/options.ts
@@ -472,13 +472,13 @@ export interface Options {
   renderSelectedChoices: 'auto' | 'always' | boolean;
 
   /**
-   * Whether selected choices should be removed from the list during search. By default selected choices appear in search results if they match the search query. Set to `false` to hide selected choices from search results.
+   * Whether selected choices should be removed from the list during search.
    *
    * **Input types affected:** select-multiple
    *
-   * @default 'auto';
+   * @default false;
    */
-  searchRenderSelectedChoices: 'auto' | 'always' | boolean;
+  searchRenderSelectedChoices: boolean;
 
   /**
    * The text that is shown whilst choices are being populated via AJAX.