+++ test/scripts/choices.test.ts
@@ -118,62 +118,6 @@ describe('choices', () => {
           });
         });
 
-        describe('passing the searchRenderSelectedChoices config option with an unexpected value', () => {
-          it('sets searchRenderSelectedChoices to true for select-multiple', () => {
-            document.body.innerHTML = `
-            <select data-choice multiple></select>
-            `;
-
-            instance = new Choices('[data-choice]', {
-              allowHTML: true,
-              searchRenderSelectedChoices: 'test' as any,
-            });
-
-            expect(instance.config.searchRenderSelectedChoices).to.equal(true);
-          });
-
-          it('sets searchRenderSelectedChoices to true for select-one', () => {
-            document.body.innerHTML = `
-            <select data-choice></select>
-            `;
-
-            instance = new Choices('[data-choice]', {
-              allowHTML: true,
-              searchRenderSelectedChoices: 'test' as any,
-            });
-
-            expect(instance.config.searchRenderSelectedChoices).to.equal(true);
-          });
-
-          it('sets searchRenderSelectedChoices to true for select-multiple', () => {
-            document.body.innerHTML = `
-            <select data-choice multiple></select>
-            `;
-
-            instance = new Choices('[data-choice]', {
-              allowHTML: true,
-              searchRenderSelectedChoices: 'test' as any,
-            });
-
-            expect(instance.config.searchRenderSelectedChoices).to.equal(true);
-          });
-        });
-
-        describe('passing the searchRenderSelectedChoices config option with "always"', () => {
-          it('sets searchRenderSelectedChoices to true', () => {
-            document.body.innerHTML = `
-            <select data-choice multiple></select>
-            `;
-
-            instance = new Choices('[data-choice]', {
-              allowHTML: true,
-              searchRenderSelectedChoices: 'always',
-            });
-
-            expect(instance.config.searchRenderSelectedChoices).to.equal(true);
-          });
-        });
-
         describe('passing the searchRenderSelectedChoices config option with false', () => {
           it('keeps searchRenderSelectedChoices as false', () => {
             document.body.innerHTML = `